Decouple SignUpWrapper from Electron
Don't require electronInterface in SignUpWrapper so that SignUpWrapper
can run in the browser. Expose function props
in SignUpWrapper so that the parent can supply the functionality that
was provided by electronInterface.
Add SignUpElectronWrapper, a component built on top of SignUpWrapper
that passes electronInterface methods as props.
Improve error handling in LoginWrapper.js